+++ src/test/java/ch/sourcepond/integrationtest/utils/TestProject.java
@@ -4,21 +4,50 @@ import static ch.sourcepond.integrationtest.utils.Photocopier.copyTestProjectToT
 
 import java.io.File;
 import java.io.FileFilter;
+import java.io.FileInputStream;
 import java.io.IOException;
+import java.io.InputStream;
 import java.util.List;
 import java.util.UUID;
 import java.util.concurrent.atomic.AtomicInteger;
+
+import javax.xml.parsers.DocumentBuilder;
+import javax.xml.parsers.DocumentBuilderFactory;
+import javax.xml.parsers.ParserConfigurationException;
+import javax.xml.xpath.XPath;
+import javax.xml.xpath.XPathExpressionException;
+import javax.xml.xpath.XPathFactory;
 
 import org.apache.commons.io.FileUtils;
 import org.apache.commons.io.filefilter.FileFilterUtils;
 import org.eclipse.jgit.api.Git;
 import org.eclipse.jgit.api.InitCommand;
 import org.eclipse.jgit.api.errors.GitAPIException;
+import org.w3c.dom.Document;
+import org.xml.sax.SAXException;
 
 import ch.sourcepond.integrationtest.E2ETest;
 
 public class TestProject {
-	public static final String PLUGIN_VERSION_FOR_TESTS = "1.0-SNAPSHOT";
+	public static final String PLUGIN_VERSION_FOR_TESTS;
+	
+	static {
+		final DocumentBuilderFactory factory = DocumentBuilderFactory.newInstance();
+		try {
+			final DocumentBuilder builder = factory.newDocumentBuilder();
+			try (final InputStream in = new FileInputStream("./pom.xml")) {
+				final Document doc = builder.parse(in);
+				final XPathFactory xpathFactory = XPathFactory.newInstance();
+				final XPath xpath = xpathFactory.newXPath();
+				PLUGIN_VERSION_FOR_TESTS = xpath.evaluate("/project/version", doc);
+			} catch (SAXException | IOException | XPathExpressionException e) {
+				throw new ExceptionInInitializerError(e);
+			}
+		} catch (ParserConfigurationException e) {
+			throw new ExceptionInInitializerError(e);
+		}
+	}
+	
 	public final File originDir;
 	public final Git origin;