+++ src/site/markdown/changelog.md
@@ -1,6 +1,11 @@
 Changelog
 ---------
 
+## 2.0.0
+ * Upgraded dependencies
+ * Maven 3.3.9 and Java 8 are required now
+ * Evaluate version of plugin to use in tests automatically
+
 ### 1.0.1
 
 * Implemented proper transfer of system properties to Maven release build